# api/API.yaml
openapi: 3.0.3
info:
  title: Car Marketplace API
  version: 1.0.0
  description: |
    - **Market Price Analysis**: Calculate the average price of cars based on brand, model, year, and location.
    - **Filtered Search**: Allow users to search for cars based on specific criteria(make, model, year, price range, mileage, fuel type, location).
    - **Geographical Market Insights**: Analyze car availability and price variations across different regions.
    - **Seller Management & Profiling**: Distinguish between professional dealers and private sellers.
    - **Buyer–Seller Communication**: Enable real-time messaging between the buyer and the seller of a specific listing.
    - **Visitor & User Registration**: Track and register platform visitors, registered users can save favourite listings.
    - **Auction Module**: Allow sellers to list cars under an auction format.
    - **Listing details and comparison**: Provide detailed information on each listing and allow users to compare multiple listings side by side based on price, mileage, year and location.
...
paths:
...
  "/api/chat":
    get:
      tags:
      - Chat
      summary: Get user chats
      operationId: getChats
      security:
      - bearerAuth: []
      responses:
        '200':
          description: User chats.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/GetChatsResponse"
        '401':
          "$ref": "#/components/responses/UnauthorizedText"
        '405':
          "$ref": "#/components/responses/MethodNotAllowedText"
        '500':
          "$ref": "#/components/responses/InternalServerErrorText"
        '502':
          "$ref": "#/components/responses/BadGatewayText"
  "/api/chat/open":
    post:
      tags:
      - Chat
      summary: Open or create a buyer-seller chat session
      operationId: openChat
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/OpenChatRequest"
      responses:
        '200':
          description: Chat session opened.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/OpenChatResponse"
        '400':
          "$ref": "#/components/responses/BadRequestText"
        '401':
          "$ref": "#/components/responses/UnauthorizedText"
        '403':
          "$ref": "#/components/responses/ForbiddenText"
        '404':
          "$ref": "#/components/responses/NotFoundText"
        '405':
          "$ref": "#/components/responses/MethodNotAllowedText"
        '500':
          "$ref": "#/components/responses/InternalServerErrorText"
        '502':
          "$ref": "#/components/responses/BadGatewayText"
  "/api/chat/{chat_id}":
    get:
      tags:
      - Chat
      summary: Get chat history
      operationId: getChatHistory
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/ChatIdPath"
      responses:
        '200':
          description: Chat history.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/GetChatHistoryResponse"
        '400':
          "$ref": "#/components/responses/BadRequestText"
        '401':
          "$ref": "#/components/responses/UnauthorizedText"
        '403':
          "$ref": "#/components/responses/ForbiddenText"
        '405':
          "$ref": "#/components/responses/MethodNotAllowedText"
        '500':
          "$ref": "#/components/responses/InternalServerErrorText"
        '502':
          "$ref": "#/components/responses/BadGatewayText"
...
  "/api/auctions":
    get:
      tags:
      - Auction
      summary: List auctions
      operationId: listAuctions
      parameters:
      - "$ref": "#/components/parameters/AuctionStatusQuery"
      - "$ref": "#/components/parameters/PageQuery"
      - "$ref": "#/components/parameters/PageSizeSnakeQuery"
      responses:
        '200':
          description: Auctions.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ListAuctionsResponse"
        '405':
          "$ref": "#/components/responses/MethodNotAllowedText"
        '500':
          "$ref": "#/components/responses/InternalServerErrorText"
        '502': &1
          "$ref": "#/components/responses/BadGatewayText"
    post:
      tags:
      - Auction
      summary: Create an auction
      operationId: createAuction
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateAuctionRequest"
      responses:
        '201':
          description: Auction created.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CreateAuctionResponse"
        '400':
          "$ref": "#/components/responses/BadRequestText"
        '401':
          "$ref": "#/components/responses/UnauthorizedText"
        '403':
          "$ref": "#/components/responses/ForbiddenText"
        '404':
          "$ref": "#/components/responses/NotFoundText"
        '405':
          "$ref": "#/components/responses/MethodNotAllowedText"
        '409':
          "$ref": "#/components/responses/ConflictText"
        '500':
          "$ref": "#/components/responses/InternalServerErrorText"
        '502':
          "$ref": "#/components/responses/BadGatewayText"
  "/api/auctions/{auction_id}":
    get:
      tags:
      - Auction
      summary: Get auction details
      operationId: getAuction
      parameters:
      - "$ref": "#/components/parameters/AuctionIdPath"
      responses:
        '200':
          description: Auction details.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/GetAuctionResponse"
        '400':
          "$ref": "#/components/responses/BadRequestText"
        '404':
          "$ref": "#/components/responses/NotFoundText"
        '405':
          "$ref": "#/components/responses/MethodNotAllowedText"
        '500':
          "$ref": "#/components/responses/InternalServerErrorText"
        '502': *1
    delete:
      tags:
      - Auction
      summary: Delete an auction
      operationId: deleteAuction
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/AuctionIdPath"
      responses:
        '204':
          description: Auction deleted.
        '400':
          "$ref": "#/components/responses/BadRequestText"
        '401':
          "$ref": "#/components/responses/UnauthorizedText"
        '403':
          "$ref": "#/components/responses/ForbiddenText"
        '404':
          "$ref": "#/components/responses/NotFoundText"
        '405':
          "$ref": "#/components/responses/MethodNotAllowedText"
        '500':
          "$ref": "#/components/responses/InternalServerErrorText"
        '502': *1
  "/api/auctions/{auction_id}/bid":
    post:
      tags:
      - Auction
      summary: Place a bid
      operationId: placeBid
      security:
      - bearerAuth: []
      parameters:
      - "$ref": "#/components/parameters/AuctionIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/PlaceBidRequest"
      responses:
        '200':
          description: Bid placed.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/PlaceBidResponse"
        '400':
          "$ref": "#/components/responses/BadRequestText"
        '401':
          "$ref": "#/components/responses/UnauthorizedText"
        '404':
          "$ref": "#/components/responses/NotFoundText"
        '405':
          "$ref": "#/components/responses/MethodNotAllowedText"
        '500':
          "$ref": "#/components/responses/InternalServerErrorText"
        '502': *1
  "/api/auctions/{auction_id}/bids":
    get:
      tags:
      - Auction
      summary: Get auction bids
      operationId: getAuctionBids
      parameters:
      - "$ref": "#/components/parameters/AuctionIdPath"
      - "$ref": "#/components/parameters/PageQuery"
      - "$ref": "#/components/parameters/PageSizeSnakeQuery"
      responses:
        '200':
          description: Auction bids.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/GetAuctionBidsResponse"
        '400':
          "$ref": "#/components/responses/BadRequestText"
        '404':
          "$ref": "#/components/responses/NotFoundText"
        '405':
          "$ref": "#/components/responses/MethodNotAllowedText"
        '500':
          "$ref": "#/components/responses/InternalServerErrorText"
        '502': *1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
  parameters:
    AuctionIdPath:
      name: auction_id
      in: path
      required: true
      schema:
        type: string
    ChatIdPath:
      name: chat_id
      in: path
      required: true
      schema:
        type: string
...
    AuctionStatusQuery:
      name: status
      in: query
      required: false
      schema:
        type: string
...
    PageQuery:
      name: page
      in: query
      required: false
      schema:
        type: integer
        format: int32
        default: 1
...
    PageSizeSnakeQuery:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        format: int32
        default: 20
...
  responses:
    BadGatewayText:
      description: Upstream WebSocket connection failed.
      content:
        text/plain:
          schema:
            type: string
    BadRequestText:
      description: Bad request.
      content:
        text/plain:
          schema:
            type: string
    ConflictText:
      description: Conflict.
      content:
        text/plain:
          schema:
            type: string
    ForbiddenText:
      description: Forbidden.
      content:
        text/plain:
          schema:
            type: string
    InternalServerErrorText:
      description: Internal server error.
      content:
        text/plain:
          schema:
            type: string
    MethodNotAllowedText:
      description: Method not allowed.
      content:
        text/plain:
          schema:
            type: string
            example: Method not allowed
    NotFoundText:
      description: Not found.
      content:
        text/plain:
          schema:
            type: string
            example: Not found
    UnauthorizedText:
      description: Unauthorized.
      content:
        text/plain:
          schema:
            type: string
            example: Unauthorized
  schemas:
...
    Auction:
      type: object
      properties:
        auction_id:
          type: string
        listing_id:
          type: integer
          format: int64
        seller_id:
          type: integer
          format: int64
        starting_price:
          type: number
          format: double
        current_price:
          type: number
          format: double
          nullable: true
        status:
          type: string
        end_time:
          type: string
        winner_user_id:
          type: string
          nullable: true
        created_at:
          type: string
        reserve_met:
          type: boolean
        total_bids:
          type: integer
          format: int32
...
    Bid:
      type: object
      properties:
        bid_id:
          type: string
        auction_id:
          type: string
        bidder_id:
          type: string
        bid_amount:
          type: number
          format: double
        timestamp:
          type: string
...
    ChatMessage:
      type: object
      properties:
        sender_id:
          type: integer
          format: int64
        content:
          type: string
        timestamp:
          type: integer
          format: int64
    ChatsSummary:
      type: object
      properties:
        chat_id:
          type: string
        listing_id:
          type: integer
          format: int64
        brand:
          type: string
        model:
          type: string
...
    CreateAuctionRequest:
      type: object
      properties:
        listing_id:
          type: integer
          format: int64
        seller_id:
          type: integer
          format: int64
          readOnly: true
          description: Overwritten from the authenticated user by the gateway.
        starting_price:
          type: number
          format: double
        end_time:
          type: string
        reserve_price:
          type: number
          format: double
          nullable: true
    CreateAuctionResponse:
      type: object
      properties:
        auction:
          "$ref": "#/components/schemas/Auction"
...
    GetAuctionBidsResponse:
      type: object
      properties:
        bids:
          type: array
          items:
            "$ref": "#/components/schemas/Bid"
        total:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
        page_size:
          type: integer
          format: int32
    GetAuctionResponse:
      type: object
      properties:
        auction:
          "$ref": "#/components/schemas/Auction"
    GetChatHistoryResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            "$ref": "#/components/schemas/ChatMessage"
        pagination:
          "$ref": "#/components/schemas/Pagination"
    GetChatsResponse:
      type: object
      properties:
        chats:
          type: array
          items:
            "$ref": "#/components/schemas/ChatsSummary"
    ListAuctionsResponse:
      type: object
      properties:
        auctions:
          type: array
          items:
            "$ref": "#/components/schemas/Auction"
        total:
          type: integer
          format: int32
        page:
          type: integer
          format: int32
        page_size:
          type: integer
          format: int32
...
    OpenChatRequest:
      type: object
      properties:
        user_id:
          type: integer
          format: int64
          readOnly: true
          description: Overwritten from the authenticated user by the gateway.
        seller_id:
          type: integer
          format: int64
        listing_id:
          type: integer
          format: int64
    OpenChatResponse:
      type: object
      properties:
        chat_id:
          type: string
        is_chat_closed:
          type: boolean
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          format: int32
        skip:
          type: integer
          format: int32
        has_next:
          type: boolean
        next_skip:
          type: integer
          format: int32
          nullable: true
    PlaceBidRequest:
      type: object
      properties:
        auction_id:
          type: string
          readOnly: true
          description: Overwritten from the path by the gateway.
        bidder_id:
          type: string
          readOnly: true
          description: Overwritten from the authenticated user by the gateway.
        bid_amount:
          type: number
          format: double
    PlaceBidResponse:
      type: object
      properties:
        bid:
          "$ref": "#/components/schemas/Bid"
...